:recycle: Refactor apache-httpclient-util module structure and dependencies
The apache-httpclient-util module has been refactored to enhance its maintainability and extendability. The `CookieStorePersistor.kt` and `NetscapeTxtCookieStorePersistor.kt` files were renamed and moved from `apache-httpclient-util-api` and `apache-httpclient-util-impl` to `apache-httpclient-util` respectively. The documentation in `CookieStorePersistor.kt` has been updated to indicate that if another implementation for a different format like JSON is needed, a new Gradle module should be created for it.

The dependencies in the build.gradle.kts files of `apache-httpclient-util-api` and `nico2vocadb-songlist-exporter` modules have been updated accordingly. The `apache-httpclient-util-api` module now includes a dependency on `com.github.CXwudi:kotlin-jvm-inline-logging`.

In the `nico2vocadb-songlist-exporter` module, the dependency on `apache-httpclient-util-impl` has been replaced with `apache-httpclient-util`.

Additionally, a minor update was made in `NetscapeTxtCookieStorePersistor.kt` to skip blank lines when reading the txt file.