:sparkles: **Add infinispan-util module and refactor cache configuration**
This commit introduces a new utility module, `infinispan-util`, which provides common configurations for Infinispan cache. Two utility functions, `setupPersist` and `setupPersistLocation`, are added to `InfinispanConfigUtil.kt`. These functions encapsulate configurations related to cache persistence and its location.

In addition, the commit refactors `CacheConfig.kt` in the `nico2vocadb-songlist-exporter` application. It now uses the newly introduced utility functions, which simplifies the code and reduces duplication.

The dependencies of `nico2vocadb-songlist-exporter` are also updated. The Infinispan core libraries are moved to `infinispan-util`, and `nico2vocadb-songlist-exporter` now depends on the `infinispan-util` module. This change aligns with our aim to have all Infinispan-related configurations and dependencies in one place.