:recycle: Refactor HTTP Configurations and Add Validation
This commit refactors the HTTP related configurations in the application. The `SystemConfigProperties` class has been renamed to `HttpConfigProperties` to better reflect its purpose. The configurations in application.yml have been updated accordingly. The existing configurations related to the HTTP client, such as `user-agent`, `cookie-jar-txt`, and `niconico-user-session-cookie-value`, are now nested under `system.http`.

In addition to the refactor, a new validation rule has been added to the `PreferenceImpl` class. Now it checks if `sortKey` and `sortOrder` are either both null or both not null.

The corresponding changes have also been applied to the test configurations and the related classes. For instance, `UserAgentRestClientCustomizer`, `NicoNicoHttpConfig`, and `PreferenceImpl` have been updated to use the newly introduced `HttpConfigProperties`.

The goal of this refactor is to make the HTTP configurations more organized and to improve the robustness of the application by ensuring the consistency of the sorting parameters.