:sparkles: Added Spring Boot annotations
Added `@SpringBootApplication` and `@ConfigurationPropertiesScan` annotations to `Nico2VocaDbSongListExporter` class in `Main.kt` file.

These annotations are essential for setting up a Spring Boot application. The `@SpringBootApplication` annotation is a convenient annotation that adds all the following:

- `@Configuration`: Tags the class as a source of bean definitions for the application context.
- `@EnableAutoConfiguration`: Tells Spring Boot to start adding beans based on classpath settings, other beans, and various property settings.
- `@ComponentScan`: Tells Spring to look for other components, configurations, and services in the `mikufan.cx.vtool.app.n2vex` package, allowing it to find and register the `MainExporterWithLocalWrite` bean.

The `@ConfigurationPropertiesScan` annotation enables support for `@ConfigurationProperties` binding on Spring Beans in the application's bean factory. This means it enables the application to bind external configurations (such as from `application.properties` or `application.yml`) to our POJOs. This makes it easier to manage and validate configurations, especially when dealing with complex hierarchical configurations.