:sparkles: **Add UserAgentRestClientCustomizer and UserAgentConfig**
This commit introduces two significant changes to the codebase:

Firstly, a new class, `UserAgentRestClientCustomizer`, has been created in the `http-service-impl` component. This class inherits from `RestClientCustomizer` and allows for the customization of the `RestClient` used in the application. The primary purpose of this class is to set a default `User-Agent` header on all HTTP requests made by the `RestClient`.

Secondly, a new `@Configuration` class, `UserAgentConfig`, has been created in the `nico2vocadb-songlist-exporter` application. This class is responsible for creating a `UserAgentRestClientCustomizer` bean with the `userAgent` property from `HttpConfigProperties`. Hence, the `User-Agent` header on all HTTP requests will be configured according to the `userAgent` property specified in the application's configuration properties.

These changes will help ensure that all HTTP requests made by the application include a `User-Agent` header, which is often required by servers to process requests.