:sparkles: Add HTTP Client Configuration
This commit adds HTTP Client configuration for the VocaDB Song List Importer in the JVM toolkit.

Three new files have been added: `HttpClientConfig.kt`, `HttpConfig.kt` and changes in `HttpConfigProperties.kt`. The `HttpClientConfig.kt` class provides the configuration for the HTTP client, including the client builder and cookie store. It also includes a `PreDestroy` method for persisting the cookie store.

The `HttpConfig.kt` file includes the configuration for the HttpServiceProxyFactory bean, RestClient customization, and the creation of the VocaDbSongListApi bean.

In `HttpConfigProperties.kt`, added a condition to check if the cookie jar text file is readable when the system requires logon. This will help to ensure that the system does not attempt to read an unreadable file. A previous configuration class `UserAgentConfig` has been removed, and its functionality has been moved to `HttpConfig.kt`.