:recycle: Refactored IOConfigImpl and FileReadableValidator
This commit introduces several changes to the IOConfigImpl.kt file. It began with updating the import statements to include `kotlin.io.path.isReadable` and `kotlin.io.path.isRegularFile`.

In the IOConfigImpl data class, the annotation `@FileReadable` was moved from the field `inputCsv` to its getter method, thus changing the target of the annotation from the field to the getter. This change allows the annotation to be applied to the returned value of the getter method, instead of the field itself.

The `@FileReadable` annotation was also updated. Besides changing its target, two more changes were made. The `@MustBeDocumented` annotation was added, ensuring that the `@FileReadable` annotation will be included in the Kotlin Doc. The `@Retention` attribute was updated from `AnnotationRetention.RUNTIME` to `AnnotationRetention.SOURCE`. This change means that the annotation will be discarded by the compiler and won't be present in the compiled class files.

The FileReadableValidator class, which is used to validate the `@FileReadable` annotation, was refactored. The file readability check was updated from `Files.isReadable(path)` to `path.isReadable()`, using the `isReadable()` extension function from `kotlin.io.path`. This change simplifies the code and makes it more idiomatic to Kotlin. Furthermore, the call to `addPropertyNode("inputCsv")` was removed, as it is no longer necessary after the refactoring.