:sparkles: **Add CoreComponentConfig and wire up Main in vsli module**
This commit introduces a new class, `CoreComponentConfig`, in the `vsli` module under the `config` package. This config class contains several Bean methods to instantiate necessary components for the module such as `loomDispatcher`, `jsonHandles`, `vocaDbSongListPusher`, `mainImporter`, `csvReader`, and `mainImporterWithLocalRead`.

Along with this, the `Main.kt` file has been modified to use the `MainImporterWithLocalRead` bean and run it. This change signifies the integration of the new components into the module's main execution process.

This update is a big step towards enhancing the functionality of the `vsli` module by incorporating various new components and connecting them for a more robust and efficient operation.