:test_tube: **Add integration test and test configuration for VocaDbSongListImporter**
This commit introduces an integration test for `VocaDbSongListImporter` (`VocaDbSongListImporterIT.kt`). This test validates that the `MainImporterWithLocalRead` function runs without throwing an exception.

The test utilizes `SpringExtension` and is run under the 'test' profile. This ensures that the test environment correctly mirrors the application environment.

Additionally, this commit adds a new configuration file (`application-test.yml`) for the test environment which includes settings for input/output CSV files and a VocaDb list id. It also includes configuration for HTTP cookies and debug level logging for HTTP headers. This configuration ensures the test environment can properly interact with VocaDb and its HTTP interfaces.

With these changes, we can now perform integration testing on the `VocaDbSongListImporter` functionality, which helps ensure its reliability and correctness.